+++ src/aigora/curriculum_graph/application/mapping/graph_mapper.py
