+++ src/aigora/curriculum_graph/application/mapping/mapper_errors.py
